refactor: move and rename files/imports related to `donation` features under `features/donations`
