+++ src/app/quero-doar/metodos/page.tsx
@@ -1,34 +1,10 @@
 'use client'
 
-import Link from "next/link";
-import Image from "next/image";
-import { BackButton } from "@/components/BackButton/backButton";
-import { PaypalButton } from "@/components/PaypalButton/PaypalButton";
-import { useDonationGuard } from "@/hooks/useDonationGuard";
+import { BackButton } from "@/features/donations/components/BackButton/BackButton";
+import { PaypalButton } from "@/features/donations/components/PaypalButton/PaypalButton"
+import { PaymentButton } from "@/features/donations/components/PaymentButton/PaymentButton";
+import { useDonationGuard } from "@/features/donations/hooks/useDonationGuard";
 
-function BotaoPagamento({
-  href,
-  icon,
-  text,
-  className = "",
-}: {
-  href: string;
-  icon: string;
-  text: string;
-  className?: string;
-}) {
-  return (
-    <Link
-      href={href}
-      className={`flex items-center gap-3 justify-center bg-[#FFF] h-[56px] rounded-[8px] border-2 
-      border-primary-400 px-6 font-semibold
-      hover:bg-primary-200 transition duration-300 w-full focus:outline-primary-800 active:bg-primary-400${className}`}
-    >
-      <Image src={icon} alt={`Ícone ${text}`} width={30} height={30} />
-      <span className="text-primary-800 text-[0.75rem] font-medium">{text}</span>
-    </Link>
-  );
-}
 export default function Page() {
   useDonationGuard();
   return (
@@ -55,7 +31,7 @@ export default function Page() {
               Você será redirecionado para o PayPal para finalizar a doação
             </p>
           </div>
-          <BotaoPagamento
+          <PaymentButton
             href="/quero-doar/metodos/pix"
             icon="/pix-icon.svg"
             text="Pagamento via PIX"
